refactor: 주문 상품 객체 대신 주문 객체가 사용자 정보를 가지도록 변경
